requirements.txt
requests==2.31.0
GitPython~=3.1.30
PyGithub~=2.1.1

# PyGithub=1.59.1
